test[stdlib][move-prover] Fix a verification failure and improve error reporting.
This fixes an issue with calling a public function in Libra.move for which the module invariant does not hold as a pre-condition. Verification was temporary turned off and is now on again.

This also improves error reporting, as devised during debugging the failure:

- Preconditions where not correctly identified and reported as post-condition failures (just the opposite from the previous bug)
- The called function is now displayed together with a precondition failure.

The error fixes lead to baseline updates.

Closes: #4740
Approved by: emmazzz