test[move-prover] Enabling modular specification.
This was easy after the change to eliminate redundant verification of postconditions.

This feature works as follows. With

```
spec fun foo {
    pragma opaque = true;
    ...
}
```

... a function will be verified standalone, but its body will not longer be inlined at caller site. However, the specified pre/post/invariant conditions will still be enforced at caller side.

This can be combined with turning verification off:

```
spec fun foo {
    pragma opaque = true;
    pragma verify = false;
    ...
}
```

With this, pre and post conditions are used at call sites, but the function itself will not be verified. This can be used to deal with notorious diffcult problems like loops or to temporarily work around Z3 termination problems.

As always, the pragmas can also be used at module level to effect all functions in a given module.

Closes: #4819
Approved by: junkil-park