test[logger] Remove no_struct_log feature
We had this feature as a temporary measure for some crates, like cluster test, where we did not want structured logging

However right now danger of this feature outweights benefits - with feature unification it can accidentally get enabled on some crate that build along the libra_node and it will kill structured logging for the whole binary.

As such it is safer to remove it.

This requires some minor change in cluster test crate to support this feature

Closes: #4821
Approved by: sherry-x