test[libra framework][move-prover] Refactor Roles.move and clients
This is a refactor of Roles.move and modules that use it to make
formal verification of access control policies less difficult, by
basing permissions on role_ids of addresses, instead of Privilege and
Capability objects.

The previous approach of publishing Privilege<Priv> objects and
extacting capabilities from them created several problems for the Move
Prover. The most fundamental of these is that the Move Prover cannot
prove a property unless it sees all of the code that could invalidate
that property. As currently designed, the available code is the module
containing the property and all modules in its dependencies.
The advantage of this approach is that properties that are so proved
will continue to hold even in an open blockchain with modules and
transactions that we do not trust.

More pragmatically, there were a number of other problems, ranging
from tricky verification problems that we haven't yet solved to
limitations of the verifier that we don't have time to fix.  The
refactor should eliminate the need to prove many properties and will
make others much simpler to prove.

Before the refactor, a Move function that is restricted to certain signers
has an argument of type &Capability<Priv>, which can only be generated
if there is a signer of the transaction with a published Privilege<Priv>.

The refactor replaces that &Capability argument with an &Signer
argument, and calls assert(has_..._privilege(account), <abort code>)
to abort the function if the signer does not have the required
permissions.  As it happens, has_..._privilege is always a property of
the role of the address as defined by the RoleId in Roles.move, so we
don't actually need to have or publish Privilege<Priv> objects.  To
achieve obvious, visible correspondence to our spreadsheet of
"governance roles and privileges", these functions have names
corresponding to privileges from the spreadsheet.  Several identical
functions have the same body, because some roles have multiple
privileges. (In the future, perhaps we can auto-generate verification
properties from something like that spreadsheet.)

This approach retains the flexibility of changing the addresses that
have a given role, because the RoleId implementation is unchanged.
The correspondence between roles and privileges can also be changed
easily by changing the appropriate has_..._privilege functions. Right
now, they all just check for a particular role ID, but there would be
no problem changing them to allow multiple alternative roles, or to
check for the presence of a published resource of some type, or lots
of other policies.

Closes: #4806
Approved by: sblackshear