test[language] loader cleanup
Closing on the loader: clean up and some changes.
A bit of renaming, commenting and code organization.
Also made the loader more of a transactional step.
If an error happens during loading, all the side effects
of loading the given module are removed. And though that was
not a problem, all things considered, it is a cleaner handling
of the cache.
This also adds a map for functions and types so we don't loop
through all the types and functions when resolving.
That looping is not disruptive, as it happened during loading,
but it does remove a possible problem with some bad cases and
scenarios.

Closes: #5030
Approved by: vgao1996