test[move-prover] Assumed/Asserted aborts and error codes.
This PR adds two new features around aborts specification, as outlined in the design doc:

- `aborts_if P with code` and `aborts_with code1, ..., codeN` allows to specify an error code. If `P` is true and the function aborts, it must abort with `code` otherwise a special verification error is produced. The error message contains detail information about expected and provided error code.  For examples, see [here](https://github.com/wrwg/libra/blob/errors/language/move-prover/tests/sources/functional/aborts_if_with_code.move).
- `aborts_if [assert] P` and `aborts_if [assume] P` allows to specify that the error condition should either become a precondition which needs to be verified to not happen, or an assumption. For examples, see [here](https://github.com/wrwg/libra/blob/errors/language/move-prover/tests/sources/functional/aborts_if_assume_assert.move).

Together with this PR comes a refactoring in spec_translator.rs on how conditions are translated for the different entry points of a function. The new approach uses a more systematic and declarative way how to do this which is easier to understand and modify using the new `ConditionDistribution` data type.

Closes: #5318
Approved by: emmazzz