test[Libra Framework][Move Prover] Revised Error Model & Specification Cleanup
This PR introduces a new approach to error handling throughout the framework, as previously discussed.

The general idea of the new approach is to divide errors into a part for the *category* (first 8 bits) and a part for the *reason*. The former is fairly general, small enumeration of codes. The later is a code local to a module which describes the background of the error. Examples:

- `Errors::not_published(ERESOURCE)` creates an error indicating that a specific resource in a module is not published when accessed.
- `Errors::requires_role(ETREASURY_COMPLIANCE)` creates an error indicating the calling account needs to have role of tresury compliance.

The capital constants are typically local to a module, and are interpreted in context of the module where the abort happened, as reported by the VM.

On specification level we write:

- `aborts_if P with Errors:NOT_PUBLISHED` specifies that a function aborts with the given error category.

Note that we currently cannot specify reasons (and perhaps we don't want to). Implementation is not trivial because we need to make the originating module explicit.

In order to keep error reporting consistent, this PR also pulls generic error conditions (like genesis status, presence of core addresses, and roles) into their matching modules. While this may reduce the fidelity of reason-based error reporting a bit, those conditions always act on the current signer(s), of which there are only a few, and require not much context. This leads to a significant increase in uniformity and brevity of code and specs, which (I hope) make it more readable.

As many of the aborts_if specifications had to be rewritten anyway for this change, I also cleaned them up and adjusted for our newest spec language features. Also, many missing specifications have been filled in while adapting the error codes.

Closes: #5463