test[storage] trivial: one less level of db dir
DB files living in
/opt/libra/data/common/libradb/db/libradb
/opt/libra/data/common/libradb/db/consensusdb
is confusing..

moving to
/opt/libra/data/common/db/libradb
/opt/libra/data/common/db/consensusdb

It'd look even cleaner if we remove common/db as well, but looks like cluster-test delete common keeping other useful things, and there are tests deleting the common parent of libradb and concensus db (It makes some sense to separate db files as well, for example, the swarm stores the mint key at the parent dir). Hence the change being like this.

Closes: #5521