test[move-prover] Fixing unsoundness in opaque aborts_if specs.
Shaz discovered a very unfortunate unsoundess bugs in aborts_if specs in the presence of `pragma opaque` as part of the more general `modifies` work.

As there is no `modifies` clause in Boogie opaque procedures for `$abort_flag` and `$abort_code`, it is assumed to not change. However, the `aborts_if` spec may say that it changes, creating unsoundness.

This is a kind of emergency fix for this issue, adding the modifies clause to the boogie output, before more wrong specs are created. As a result, multiple specs now not longer verify and/or start to timeout. This are commented out for now. Fixing those cases will conflict with changes in the error work, so they are happening in this PR.

Closes: #5575