test[Move Prover] Optimizations for global invariants/removal of some timeouts
This PR optimizes global invariants by skipping verification on updates if the memory used by the invariant is not part of the VC (i.e. is not from any modules on the command line, but only from dependencies).

This brings back some advantages of module invariants. Consider the case of a module invariant in M which only refers to memory of module M. This invariant is only included in verification if M itself is verified. This is sound because of data isolation in Move, as only code in M can violate this invariant. Notice that the invariant is still assumed outside of M, to express constraints over its memory.

Now consider the generalized case of a module invariant which touches memory from M1 and M2. It is included in verification if either M1 is verified, M2 is verified, or both.

Notice that the problem of modular verification and global invariants does not go away with this approach: if M1 uses M2 and imposes new constraints on memory of M2, we would still need to verify M1 and M2 together (multiple module arguments to MVP). However, the case is rare and does not currently happen in the framework AFAICT. It is generally safe to use another modules memory in the premise of an implication. For instance, we frequently have invariants of the form `LibraTimestamp::is_operating() => P`. This invariant does not impose new constraints on `LibraTimestamp` because it is referred to in a premise.

The new compilation schema removes, however, once instance of where verification happened by 'accident' before. Suppose M1 adds constraints to M2 and calls a function M2::f. Now we verify M1 standalone, without M2. Before this PR, if this function updates by luck memory in M2, the additional constraints from M1 are checked. Not so luckily if M::f does not update memory. This source of inconsistency is removed with this PR; if M1 uses M2 and imposes new constraints on M2, those need to be always verified together.

In addition, this PR renames the property `[on_update]` into `[isolated]` to better express its intention. An invariant marked as isolated does not contribute to the verification context of other conditions. It is only assumed if the underlying memory is actually updated. It is assumed before the update so that the prover can use the pre-state to maintain the invariant for the post-state.

This leads to some speed improvements and removal of a number of timeouts. There have been also functions forgotten to reactivate for verification in an older PR, specifically in LibraAccount. This is fixed with this PR as well.

Closes: #5732