test[libra framework] add some comments to FixedPoint32 about precision
Financial applications usually deal with decimal numbers, and the
binary representation used for FixedPoint32 can cause confusion.
Add some comments to explain the available precision and suggest
a guideline for avoiding trouble when using the create_from_rational
function.

This also fixes a comment about when that function aborts and removes
a note about creating a ratio of zero. Those comments go back to the
original version of this code, but they were not quite right. The
code intentionally allows creating a ratio of zero by specifying a
zero numerator.

Closes: #5669