test[breaking][network] Remove Raw(Enc)NetworkAddress
* Merge RawNetworkAddress encoding into NetworkAddress, so that we can
still handle incompatible versions in a Vec<Serialize Network Addresses>
* Eliminate RawEncNetworkAddress since versionining should both be
uniform and the format is so simple as to not benefit from versioning.

This is a breaking change because RawEncNetworkAddress is no longer
present and instead is represented on chain as a EncNetworkAddress
directly.

Removing because it came up often enough that it was confusing people
why we had this layer of abstraction... and writing the addresses is an
atomic operation, if the validator operator messed up one, then they
probably messed up the rest.

Closes: #5766