test[network] Fix peer_id naming in Peer actor
The peer_id of the actor for a Peer was confusingly named self_peer_id,
which became even more confusing of whether the peer_id was of the local
peer, or of the peer actor.  It has been renamed remote_peer_id, to more
appropriately match that it is not the current machine.

Closes: #5982