test[language] Structured logging in the VM
This is to integrate with structured logging and the rest of the system.
It should allow us to connect errors to other part of the overall system.
It implies passing a logger to the VM, where the logger is a client controlled instance.
That ends up being a very intrusive change, yet inevitable.
This is the initial PR that allows us to hook up the logger in all critical place,
more PRs will be coming shortly to add more info.
I also need to make sure everything works as expected and will do it in the next day or
two once I understand what a good way to test this in the overall system is.

Closes: #6237