test[network] Add proxy protocol to network transport and config
Now there is a field in the NetworkConfig enable_proxy_protocol, which
allows for changing proxy protocol based on the environment it is in.
In the case where a deployment strategy uses a proxy such as HAProxy in
front of the node's network interface, proxy protocol can be turned on
for both.