test[vm] fix off-by-1 error in PC reported in aborting instructions
The bytecode interpreter increments the PC before executing an instruction. This means that if the instruction aborts, the PC location associated with the ExecutionFailure is off by 1.

This PR fixes the issue by moving the increment to the end of the interpreter loop and adding PC increments when:

- pushing a new frame on when returning from a call
- returning from a native function call
This enforces the following invariant: the PC advances to instruction i+1 iff instruction i executed without aborting.

Closes: #6292