test[libra framework] remove unused Debug and Compare modules from Libra Framework
Both of these modules are useful, but not called by any current transaction scripts. This PR moves them to a nursery/ directory under stdlib/ that is not compiled or published alongside the other Libra Framework modules.
This moves the tests for Compare and Debug to Move CLI expected value tests to avoid the headache of maintaining multiple genesis blobs in the stdlib.
In addition, this tests the exact output of Debug, which is useful.

Closes: #6383