test[move-prover] Cleanup of Genesis, Hash, and LCS.
This also fixes a bug in the specs of the code called by Genesis, namely that the constructed auth key was *always* invalid leading to early abortion (for the prover) of Genesis::initialize. With this fix, it is theoretically possible to verify whether Genesis establishes the expected invariants (by calling e.g. `mvp Genesis.move LibraAccount.move`, i.e. pairwise verification of Genesis with a module providing an invariant). However, because most of the M::initialize functions are not opaque, this currently times out.

Closes: #6439