test[docgen] Update documentation generator to support multiple files.
This updates the documentation generator to support not only one root template but multiple ones. This allows us to write different markdown files for top-level documentation of different aspects of the framework which are fully processed for syntax highlighting and cross-references.

Also adds a new `./language/stdlib/transaction_scripts/spec_documentation_template.md` file which is the top-level page for overview documentation of specifications in the framework. This is a strawman which will be populated in the next few days.

*Why this change is safe*

Only code in `./language/move-prover/docgen` and `./language/stdlib/src/lib.rs` is effected. This code does not run in production.

Closes: #6467