test[language] Add trace! to failing transactions during execution
Add trace! for transactions that get discarded during execution.
When validating transactions, mempool already reports failures so
we don't need to do it in the VM or Adapter.
However execution was not reporting discarded transaction unless
coming from invariant violation. And discarded transactions are
not charged and a system cost.
Given that trace is not logged to the server and can be enabled
"on the fly" this is a good way to debug possible problems
without penalizing the system.

Closes: #6509